Update Gaia API documents for #18583c3be0357e4fa5224debf7eeb9c081ea3ff3     Bug 916850: Skip multi-page Ogg meta data
Ogg files are split into multiple segments, called pages. Currently
we only support meta data in a single page. This can result in parser
errors when meta data fields span multiple pages, and the inability
to listen to the Ogg file.

With this patch, the parser simply stops in these cases. We don't get
the respective information from the file, but the file will at least
be available for playing in the Music player.

Signed-off-by: Thomas Zimmermann <tdz@users.sourceforge.net>